CreateUploadSignature route
Adding a route to create TUS upload signatures using a provided libraryId. This is not a built-in Bunny API route, but in a real-world scenario using this project, it would be impossible to create a TUS signature otherwise without another service having access to API keys otherwise irrelevant to the service's purpose.